+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - [browser] feat: Expose `wrap` function in `@sentry/browser`
+- [browser] - feat: Added `onLoad` callback to `showReportDialog`
 
 ## 5.2.0